+++ takina/cogs/owner-utils.py
@@ -216,6 +216,11 @@ class OwnerUtils(commands.Cog):
     @commands.command(hidden=True)
     @commands.is_owner()
     async def pull(self, ctx: commands.Context):
+        if os.get_env("NIXOS_INSTANCE"):
+            embed = nextcord.Embed()
+            embed.color = config.ERROR_COLOR
+            embed.description = ":x: The pull command is not available on instances of {config.BOT_NAME} run on NixOS."
+            await ctx.reply(embed=embed, mention_author=False)
         current_dir = os.getcwd()
 
         def run_git_pull(directory):