+++ CHANGELOG.md
@@ -136,4 +136,4 @@ Added `2hu` alias for the `touhou` command and `umamusume` alias for the `uma` c
 Added `myanimelist` alias for `mal` command.
 
 ### 1.16.0
-Added `listeners.honeypot`, `revert` command to go along with the `commit` command and make Docker and the repository itself use uv.
+Added `listeners.honeypot` and make Docker and the repository itself use uv.